% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load.R
\name{download_infosiga}
\alias{download_infosiga}
\title{Download and Extract Infosiga data}
\usage{
download_infosiga(destpath)
}
\arguments{
\item{destpath}{A string specifying the directory path where the data
files should be extracted. The path will be normalized, and if it's a
relative path, it will be resolved from the current working directory.}
}
\value{
This function does not return a value. It is called for its side effects:
downloading a file and extracting it to the \code{destpath}.
}
\description{
Downloads the complete Infosiga.SP database from the official website,
handles potential HTTP errors with retries, and extracts the contents
of the resulting ZIP file to a specified directory.
}
\details{
The function performs the download to a temporary file and then unzips it.
It uses a robust \code{httr2} request that forces HTTP/1.1, sets a common
User-Agent, and automatically retries the download up to 3 times in case of
transient network errors. A progress bar is displayed during the download.
}
\examples{
\dontrun{
# Create a temporary directory to store the data
temp_dir <- tempdir()

# Download and extract the data
download_infosiga(destpath = temp_dir)
}
}
