% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{palette_detran}
\alias{palette_detran}
\title{Detran Color Palette}
\usage{
palette_detran()
}
\value{
A named list of character strings, where each string is a hex color code.
}
\description{
Provides a list of hex color codes representing the official Detran color palette.
}
