% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{theme_detran}
\alias{theme_detran}
\title{Custom ggplot2 Theme for Detran Visualizations}
\usage{
theme_detran(...)
}
\arguments{
\item{...}{Additional arguments passed to \code{ggplot2::theme()}.}
}
\value{
A \code{ggplot2} theme object.
}
\description{
Provides a custom ggplot2 theme with a minimal base, adjusted legend
positioning, and other aesthetic modifications suitable for Detran-related
data visualizations.
}
